requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-bootstrap4==3.0.1
django-rest-registration==0.6.3
python-dotenv==0.18.0
psycopg2==2.9.1
